MATBAIJ: merge two files to avoid generating an object file with no symbols
